requirements3.dev.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
coverage==4.4.1
cycler==0.10.0
decorator==4.1.2
Flask==0.12.2
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mock==2.0.0
numpy==1.13.1
obspy==1.0.3
pandas==0.20.3
pbr==3.1.1
pep8==1.7.0
psutil==5.3.1
psycopg2==2.7.3.1
py==1.4.34
pylint==1.7.2
pyparsing==2.2.0
pytest==3.2.2
pytest-cov==2.5.1
pytest-mock==1.6.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
scipy==0.19.1
six==1.10.0
SQLAlchemy==1.1.14
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
